Disable user and group setup for openvpn server.
1.Openvpn daemon need to be running as "nobody" user and "nobody" group
and regarding syscall(setuid/setgid) will trigger apparmor denies error
if we install the snap and run it in strict mode. We disable user and
group management for the time being so that we can release the snap in
stable channel.
2.In the todo list, I removed "tun/tap for openvpn...  mode" section
as read/write acccess to /dev/net/tun is allowed in snapd with
network-control connected.